"Finished"
Added some null checks when needed because sometimes the site was missing things.
Added a subpath to the cache so we can work with sites with the same file name (in this case level1/meishi.html + level2/meishi.html etc)
Gave random id numbers to model, deck, etc.
Added arguments to the code so we can parse levels 1,2,3.
Added subdecks to separate nouns, verbs, adjectives.